Create Matrix Game 1
Problem statement
Ninja is a teacher at a school. He introduced a game of matrix. He gives a square matrix, i.e., NxN matrix, to all the school students and asks them to check if the matrix is idempotent or not.

A matrix is an idempotent matrix if a matrix multiplied by itself returns the same matrix. The matrix M is said to be an idempotent matrix if and only if M * M = M. In the idempotent matrix, M is a square matrix.

Among them, a student Ninja is new to programming; he doesn’t have much experience; he asks you to solve the problem. Can you help Ninja figure out whether the matrix is idempotent?

Detailed explanation ( Input/output format, Notes, Images )
Constraints:
1<= T <= 50
1<= N <= 100 

 Time Limit: 1 sec
Sample Input 1:
2
3
2 -2 -4
-1 3 4
1 -2 -3
2
1 1
1 1
Sample Output 1:
True
False
Explanation for Sample Input 1:
In the first test case, the matrix M = [[2,-2,-4],[-1,3,4],[1,-2,-3]] which on multiplying with itself yields the same matrix M. Hence the matrix is idempotent.

In the second test case, the matrix M = [[1,1],[1,1]] which on multiplying with itself does not yield the same matrix M. Hence the matrix is not idempotent.
Sample Input 2:
2
2
3 -6
1 -2
3
1 2 3
4 5 6
7 8 9
Sample Output 2:
True
False