Create Serialize and Deserialize an N-ary tree
Problem statement
You are given an N-ary tree where every node has at most ‘N’ child nodes. You need to first serialize it and then deserialize the serialized tree.

Serialization is the process of converting a data structure or object into a sequence of bits so that it can be stored in a file or memory buffer, or transmitted across a network connection link to be reconstructed later in the same or another computer environment.

In simple words, serialization is a process to store a tree in a file so that it can be later restored to its original form. Deserialization is reading the tree back from the file.

Note:
Note that the structure of the tree must be maintained while serialization and deserialization. Also, you can use any method to serialize or deserialize the tree. You just need to see that the tree can be serialized into a string and further deserialized into the original tree.
Detailed explanation ( Input/output format, Notes, Images )
Constraints
1 <= T <= 10
1 <=  N <= 10^4

Time limit: 1 sec
Sample Input 1
1
1 2 3 4 -1 5 6 -1 -1 -1 -1 -1 
Sample Output 1
1 2 3 4 -1 5 6 -1 -1 -1 -1 -1 
Explanation
After deserialization, you need to return the same tree, so print 1 2 3 4 -1 5 6 -1 -1 -1 -1 -1 
Sample Input 2:
2
1 2 3 4 5 -1 6 7 8 -1 9 -1 -1 -1 -1 -1 -1 -1
1 2 -1 -1
Sample Output 2
1 2 3 4 5 -1 6 7 8 -1 9 -1 -1 -1 -1 -1 -1 -1
1 2 -1 -1
Explanation
In the first test case
After deserialization, you need to return the same tree, so print 1 2 3 4 5 -1 6 7 8 -1 9 -1 -1 -1 -1 -1 -1 -1 

In the second test case
After deserialization, you need to return the same tree, so print 1 2 -1 -1