Create Range Sum Query - Mutable
Problem statement
You are given an array ‘ARR’. You are supposed to process two types of queries - Update an index ‘IND’ with value ‘VAL’ in the array, and find the sum of a range in the array.

You are supposed to implement the class which includes two operations:

1. UPDATE_INDEX(IND, VAL) - It updates the value of ARR[IND] to VAL.
2. SUM\_OF\_RANGE(l, r) - It returns the sum of the subarray ARR[l] to ARR[r] i.e. ARR[l] + ARR[l+1] + ARR[l+1] + ….. + ARR[r-1] + ARR[r].
Detailed explanation ( Input/output format, Notes, Images )
Constraints:
1 <= T <= 50  
1 <= Q <= 10^4
1 <= N <= 10^4
-10^3 <= ARR[i] <= 10^3
TYPE_OF_QUERY = {0, 1}
0 <= IND <= N
-100 <= VAL <= 100
0 <= l <= r <= N

Time Limit: 1 sec
Sample Input 1:
2
6
1 2 3 4 5 6
3
1
2 5
0
3 5
1
2 5   
5
9 8 7 6 5
3
0
4 13
1
3 4
1
0 4
Sample Output 1:
18
19
19
43
Explanation for Sample Input 1:
In the first test case, The initial array is [1, 2, 3, 4, 5, 6]. In the first query, the sum from index 2 to 5 is 18. In the second query, the third index is updated from 4 to 5. In the third query, the sum from index 2 to 5 is 19.

In the second test case, the given array is [9, 8, 7, 6, 5]. In the first query, index 4 is updated from 5 to 13. In the second query, the sum from index 3 to 4 is 19. In the third query, the sum from index 0 to 4 is 43.
Sample Input 2:
2
4
1 3 5 7 
2
0 1 1 
1 1 3 
3
1 3 5 
2
0 1 1 
1 1 2 
Sample Output 2:
13
6